README.md
This Railway Management System is built in Python using an object-oriented approach. It allows adding trains, schedules, and passengers, and supports ticket booking with seat availability checks and automatic seat assignment. Data is stored persistently using JSON files, eliminating the need for a database. The system features a menu-driven command-line interface for ease of use, with built-in validation to prevent duplicates and invalid bookings. Ideal for learning Python, OOP, and file handling while simulating a real-world railway booking process.